Documentation

DyadicKDE.DyadicKernelDensityEstimator — Type

Composite type for the dyadic kernel density estimator

DyadicKDE.DyadicKernelDensityEstimator — Method

Construct a dyadic kernel density estimator

Base.Multimedia.display — Method

Display a dyadic kernel density estimator

DyadicKDE.estimate_ROT_bandwidth — Method

Estimate a rule-of-thumb bandwidth from dyadic data

DyadicKDE.fit — Method

Fit a dyadic kernel density estimator

DyadicKDE.get_RIMSE — Method

Compute the root integrated mean squared error

DyadicKDE.get_bci_average_width — Method

Return the average width of the Bonferroni confidence intervals

DyadicKDE.get_bci_coverage — Method

Check if the Bonferroni confidence intervals all cover the true density function

DyadicKDE.get_f — Method

Get the true density function from example dyadic data

DyadicKDE.get_pci_average_width — Method

Return the average width of the pointwise confidence intervals

DyadicKDE.get_pci_coverage — Method

Check if the pointwise confidence intervals all cover the true density function

DyadicKDE.get_ucb_average_width — Method

Return the average width of the uniform confidence band

DyadicKDE.get_ucb_coverage — Method

Check if the uniform confidence band covers the true density function

DyadicKDE.make_data — Method

Generate some example dyadic data using a Gaussian mixture model

DyadicKDE.mean — Method

Compute the mean of a vector of numbers

DyadicKDE.phi — Method

Compute the standard normal density function