Documentation

Dyadic kernel density estimation

DyadicKDE.DyadicKernelDensityEstimator — Type

Composite type to represent a dyadic kernel density estimator.

DyadicKDE.DyadicKernelDensityEstimator — Method

DyadicKernelDensityEstimator(kernel_name, bandwidth, significance_level,
                             n_resample, sdp_solver, evals, W, meta)

Construct a dyadic kernel density estimator.

Arguments

kernel_name::String: which kernel to use.

bandwidth::Float64: the bandwidth for the estimator.

significance_level::Float64: for the confidence band/intervals.

n_resample::Int: the number of resamples used to construct the confidence band/intervals.

sdp_solver::String: semi-definite program solver.

evals::Vector{Float64}: points at which to evaluate the density estimator.

W::UpperTriangular{Float64}: array of dyadic data.

meta::Dict: any extra information to pass to the estimator.

Base.Multimedia.display — Method

display(est::DyadicKernelDensityEstimator)

Display a dyadic kernel density estimator.

DyadicKDE.estimate_ROT_bandwidth — Method

estimate_ROT_bandwidth(W::UpperTriangular{Float64},
                       kernel_name::String)

Estimate a rule-of-thumb bandwidth from dyadic data.

DyadicKDE.fit — Method

fit(est::DyadicKernelDensityEstimator)

Fit a dyadic kernel density estimator to data.

Counterfactual dyadic kernel density estimation

DyadicKDE.CounterfactualDyadicKernelDensityEstimator — Type

Composite type to represent a counterfactual dyadic kernel density estimator.

DyadicKDE.CounterfactualDyadicKernelDensityEstimator — Method

CounterfactualDyadicKernelDensityEstimator(kernel_name, bandwidth, significance_level,
                                           n_resample, sdp_solver, evals,
                                           W, X0, X1, meta)

Construct a counterfactual dyadic kernel density estimator.

Arguments

kernel_name::String: which kernel to use.

bandwidth::Float64: the bandwidth for the estimator.

significance_level::Float64: for the confidence band/intervals.

n_resample::Int: the number of resamples used to construct the confidence band/intervals.

sdp_solver::String: semi-definite program solver.

evals::Vector{Float64}: points at which to evaluate the density estimator.

W::UpperTriangular{Float64}: array of treated dyadic data.

X0::Vector{Int}: categorical vector of untreated covariates.

X1::Vector{Int}: categorical vector of treated covariates.

meta::Dict: any extra information to pass to the estimator.

DyadicKDE.ParametricCounterfactualDyadicKernelDensityEstimator — Type

Composite type to represent a counterfactual dyadic kernel density estimator.

DyadicKDE.ParametricCounterfactualDyadicKernelDensityEstimator — Method

ParametricCounterfactualDyadicKernelDensityEstimator(kernel_name, bandwidth, significance_level,
                                                     n_resample, sdp_solver, evals,
                                                     W, X0, X1, meta)

Construct a parametric counterfactual dyadic kernel density estimator.

Arguments

kernel_name::String: which kernel to use.

bandwidth::Float64: the bandwidth for the estimator.

significance_level::Float64: for the confidence band/intervals.

n_resample::Int: the number of resamples used to construct the confidence band/intervals.

sdp_solver::String: semi-definite program solver.

evals::Vector{Float64}: points at which to evaluate the density estimator.

W::UpperTriangular{Float64}: array of treated dyadic data.

X0::Vector{Int}: categorical vector of untreated covariates.

X1::Vector{Int}: categorical vector of treated covariates.

phat0::Vector{Float64}: vector of precomputed parametric untreated covariate estimates.

phat1::Vector{Float64}: vector of precomputed parametric treated covariate estimates.

meta::Dict: any extra information to pass to the estimator.

Base.Multimedia.display — Method

display(est::CounterfactualDyadicKernelDensityEstimator)

Display a counterfactual dyadic kernel density estimator.

Base.Multimedia.display — Method

display(est::ParametricCounterfactualDyadicKernelDensityEstimator)

Display a parametric counterfactual dyadic kernel density estimator.

DyadicKDE.fit — Method

fit(est::CounterfactualDyadicKernelDensityEstimator)

Fit a counterfactual dyadic kernel density estimator to data.

DyadicKDE.fit — Method

fit(est::ParametricCounterfactualDyadicKernelDensityEstimator)

Fit a parametric counterfactual dyadic kernel density estimator to data.

Helper functions

DyadicKDE.get_RIMSE — Method

get_RIMSE(fhat::Vector{Float64}, f::Vector{Float64})

Compute the root integrated mean squared error of an estimate for a function.

DyadicKDE.get_average_width — Method

get_average_width(cb::Matrix{Float64})

Return the average width of a confidence band.

DyadicKDE.get_coverage — Method

get_coverage(cb::Matrix{Float64}, f::Vector{Float64})

Check if a confidence band covers the true density function.

DyadicKDE.get_f — Method

get_f(p::Vector{Float64}, evals::Vector{Float64})

Get the true density function from example dyadic Gaussian mixture data.

DyadicKDE.make_dyadic_data — Method

make_dyadic_data(n_data::Int, p::Vector{Float64})

Generate some example dyadic data using a Gaussian mixture model.

DyadicKDE.mean — Method

mean(x::Vector{<:Number})

Compute the mean of a vector of numbers.

DyadicKDE.phi — Method

phi(t::Real)

Compute the standard normal density function.