Documentation

Dyadic kernel density estimation

DyadicKDE.DyadicKernelDensityEstimator — Type

Composite type to represent a dyadic kernel density estimator.

DyadicKDE.DyadicKernelDensityEstimator — Method

DyadicKernelDensityEstimator(kernel_name, bandwidth, significance_level,
                             n_resample, sdp_solver, evals, data, meta)

Construct a dyadic kernel density estimator.

Arguments

kernel_name::String: which kernel to use.

bandwidth::Float64: the bandwidth for the estimator.

significance_level::Float64: for the confidence band/intervals.

n_resample::Int: the number of resamples used to construct the confidence band/intervals.

sdp_solver::String: semi-definite program solver.

evals::Vector{Float64}: points at which to evaluate the density estimator.

data::UpperTriangular{Float64}: array of dyadic data.

meta::Dict: any extra information to pass to the estimator.

Base.Multimedia.display — Method

display(est::DyadicKernelDensityEstimator)

Display a dyadic kernel density estimator.

DyadicKDE.estimate_ROT_bandwidth — Method

estimate_ROT_bandwidth(data::UpperTriangular{Float64},
                       kernel_name::String)

Estimate a rule-of-thumb bandwidth from dyadic data.

DyadicKDE.fit — Method

fit(est::DyadicKernelDensityEstimator)

Fit a dyadic kernel density estimator to data.

Helper functions

DyadicKDE.get_RIMSE — Method

get_RIMSE(est::DyadicKernelDensityEstimator)

Compute the root integrated mean squared error.

DyadicKDE.get_bci_average_width — Method

get_bci_average_width(est::DyadicKernelDensityEstimator)

Return the average width of the Bonferroni confidence intervals.

DyadicKDE.get_bci_coverage — Method

get_bci_coverage(est::DyadicKernelDensityEstimator)

Check if the Bonferroni confidence intervals all cover the true density function.

DyadicKDE.get_f — Method

get_f(est::DyadicKernelDensityEstimator)

Get the true density function from example dyadic Gaussian mixture data.

DyadicKDE.get_pci_average_width — Method

get_pci_average_width(est::DyadicKernelDensityEstimator)

Return the average width of the pointwise confidence intervals.

DyadicKDE.get_pci_coverage — Method

get_pci_coverage(est::DyadicKernelDensityEstimator)

Check if the pointwise confidence intervals all cover the true density function.

DyadicKDE.get_ucb_average_width — Method

get_ucb_average_width(est::DyadicKernelDensityEstimator)

Return the average width of the uniform confidence band.

DyadicKDE.get_ucb_coverage — Method

get_ucb_coverage(est::DyadicKernelDensityEstimator)

Check if the uniform confidence band covers the true density function.

DyadicKDE.make_data — Method

make_data(n_data::Int, p::Vector{Float64})

Generate some example dyadic data using a Gaussian mixture model.

DyadicKDE.mean — Method

mean(x::Vector{<:Number})

Compute the mean of a vector of numbers.

DyadicKDE.phi — Method

phi(t::Real)

Compute the standard normal density function.